# New convex hull for Polygon2D