… src/Spatial/Internals/ConvexHull/ConvexHull.cs
… src/Spatial/Internals/ConvexHull/QuadrantSpecific2.cs
… src/Spatial/Internals/ConvexHull/QuadrantSpecific3.cs
… src/Spatial/Internals/ConvexHull/QuadrantSpecific4.cs
… src/Spatial/Spatial.csproj
… src/SpatialUnitTests/Euclidean/Polygon2DTests.cs