Implements a single threated version of Ouellet's algorithm O(n log h) for convex hull calculation this is based upon the work of Lui and Chen (2007) and replaces the Quickhull method previously used.
Liu, G. & Chen, C. J. Zhejiang Univ. - Sci. A (2007) 8: 1210. https://doi.org/10.1631/jzus.2007.A1210 

This provides a fix #59 

Also changed assessors for Vertices and Edges to be enumerators.