@AQuentin2033 2017-12-09
Performance test results:


Method | Current Mean   (us) | Current Allocated   (KB) | Proposed Mean   (us) | Proposed Allocated   (KB) | Performance Increase | Mem Use Reduction
-- | -- | -- | -- | -- | -- | --
GetConvexHullFromPoints10 | 5.741 | 3.5 | 3.346 | 2.76 | 172% | 127%
GetConvexHullFromPoints100 | 59.215 | 18.97 | 9.914 | 8.84 | 597% | 215%
GetConvexHullFromPoints1000 | 674.968 | 121.34 | 68.949 | 52.5 | 979% | 231%
GetConvexHullFromPoints10000 | 20,893.22 | 1108.31 | 615.675 | 675.06 | 3394% | 164%